requirements.txt
# App Requiremments

transformers[accelerate]
prompt_toolkit
termcolor
setuptools
bitsandbytes
protobuf
sentencepiece
pynvml
pysbd
sentence_transformers
mysql-connector-python
qdrant-client
python_dotenv
vllm
requests
click
mistral_common
mistral_inference


# Dev Requirements
huggingface_hub[cli]
wcwidth
matplotlib
